Move UiState and UiEvent from AllFavoritesScreen to AllFavoritesPresenter
