+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/home/AllFavoritesLayout.kt
@@ -24,8 +24,8 @@ import org.burnoutcrew.reorderable.detectReorderAfterLongPress
 import org.burnoutcrew.reorderable.rememberReorderableLazyListState
 import org.burnoutcrew.reorderable.reorderable
 import org.cru.godtools.R
-import org.cru.godtools.ui.dashboard.home.AllFavoritesScreen.UiEvent
-import org.cru.godtools.ui.dashboard.home.AllFavoritesScreen.UiState
+import org.cru.godtools.ui.dashboard.home.AllFavoritesPresenter.UiEvent
+import org.cru.godtools.ui.dashboard.home.AllFavoritesPresenter.UiState
 import org.cru.godtools.ui.tools.ToolCard
 
 @Composable
