+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/home/AllFavoritesPresenter.kt
@@ -11,6 +11,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import com.slack.circuit.codegen.annotations.CircuitInject
+import com.slack.circuit.runtime.CircuitUiEvent
+import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.presenter.Presenter
 import com.slack.circuitx.android.IntentScreen
@@ -27,8 +29,7 @@ import org.cru.godtools.analytics.model.OpenAnalyticsActionEvent.Companion.ACTIO
 import org.cru.godtools.analytics.model.OpenAnalyticsActionEvent.Companion.SOURCE_FAVORITE
 import org.cru.godtools.db.repository.ToolsRepository
 import org.cru.godtools.model.Tool
-import org.cru.godtools.ui.dashboard.home.AllFavoritesScreen.UiEvent
-import org.cru.godtools.ui.dashboard.home.AllFavoritesScreen.UiState
+import org.cru.godtools.ui.dashboard.home.AllFavoritesPresenter.UiState
 import org.cru.godtools.ui.tooldetails.ToolDetailsScreen
 import org.cru.godtools.ui.tools.ToolCard
 import org.cru.godtools.ui.tools.ToolCardPresenter
@@ -44,6 +45,17 @@ class AllFavoritesPresenter @AssistedInject constructor(
     @Assisted
     private val navigator: Navigator,
 ) : Presenter<UiState> {
+    @ConsistentCopyVisibility
+    data class UiState internal constructor(
+        val tools: List<ToolCard.State> = emptyList(),
+        internal val eventSink: (UiEvent) -> Unit = {},
+    ) : CircuitUiState
+
+    internal sealed interface UiEvent : CircuitUiEvent {
+        data class MoveTool(val from: Int, val to: Int) : UiEvent
+        data object CommitToolOrder : UiEvent
+    }
+
     @Composable
     override fun present(): UiState {
         val scope = rememberCoroutineScope()
